feat(KUI-1462): Refactor use of analysis component
